pyproject.toml
[tool.poetry]
name = "point-cloud-attention"
version = "0.1.0"
description = ""
authors = ["Cristian Garcia <cgarcia.e88@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.1"
tensorflow = "^2.1.0"
typer = "^0.0.8"
pyyaml = "^5.3"
cytoolz = "^0.10.1"
scikit-learn = "^0.22.1"
plotly = "^4.5.0"
matplotlib = "^3.1.3"
dataget = "^0.4.1"
elegy = {path = "../elegy"}

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
